feat: add MunicipalityController.acceptEvent and its proper execution flow

+++ src/main/BusinessLogic/MunicipalityController.java
@@ -102,28 +102,40 @@ public class MunicipalityController extends BasicUserController implements Obser
     }
 
     private void acceptEvent() throws SQLException {
-        // print events to be accepted
+        // Create object to be removed and the list that will be used
+        PublicEvent toBeRemoved = null;
         ArrayList<PublicEvent> events = municipality.getEventsToBeAccepted();
-        municipalityInterface.printPublicEvents(events);
 
-        // take input as id of the event
-        System.out.println("Select the ID of the event you want to accept: ");
-        input = getInteger();
+        // Checks if the array is not empty
+        if (!events.isEmpty()) {
+            // print events to be accepted
+            municipalityInterface.printPublicEvents(events);
 
-        // search for the requested event
-        if (input >= 0) {
-            for (PublicEvent event: events){
-                if (event.getId() == input){
-                    // set the accepted field to true
-                    event.setAccepted(TRUE);
-                    // remove event from array
-                    municipality.getEventsToBeAccepted().remove(event);
-                    // remove event from table in database
-                    EventsToBeAcceptedDAO.delete(event.getId());
+            // take input as id of the event
+            System.out.println("---------------");
+            System.out.println("Select the ID of the event you want to accept: ");
+            input = getInteger();
+
+            // search for the requested event
+            if (input >= 0) {
+                for (PublicEvent event: events){
+                    if (event.getId() == input){
+                        toBeRemoved = event;
+                        // set the accepted field to true
+                        event.setAccepted(TRUE);
+                        // remove event from table in database
+                        EventsToBeAcceptedDAO.delete(event.getId());
+                        // Show results
+                        System.out.println("Event ACCEPTED!");
+                    }
                 }
+                // remove event from array
+                municipality.delete_event(toBeRemoved);
+            } else { // Error message if the ID is not valid
+                accessInterface.invalidChoice();
             }
-        } else {
-            accessInterface.invalidChoice();
+        } else { // Here the array is empty
+            System.out.print("No Events to be accepted. ");
         }
     }
